Outsourced the offset_slope calculations
-for better overview
- also added some comments for better understanding
